[+] changes in deployment script

+++ src/contracts/ProceedsPayments.sol
@@ -2,8 +2,10 @@
 
 pragma solidity ^0.8.0;
 
-import "@openzeppelin/contracts/finance/PaymentSplitter.sol";
+import '@openzeppelin/contracts/finance/PaymentSplitter.sol';
 
 contract ProceedsPayments is PaymentSplitter {
-    constructor(address[] memory payees_, uint[] memory shares_) PaymentSplitter(payees_, shares_) {}
+    constructor(address[] memory payees_, uint256[] memory shares_)
+        PaymentSplitter(payees_, shares_)
+    {}
 }
